Agregué el módulo graphics, para no mezclar la lógica de el módulo board con la interfaz
