+++ app/src/main/java/com/modules/logic/Board.java
@@ -35,8 +35,12 @@ class Board{
         return board[yCoordinate][xCoordinate].getSymbol() == ' ';
     }
 
-
-
+    /**
+     * This function will return ithe symbol in the coordenates given
+     */
+    private char getBoardSymbol(int yCoordinate, int xCoordinate){ //verifies if the symbol placement is valid.
+        return board[yCoordinate][xCoordinate].getSymbol();
+    }
 
     /**
      * This function will determine if the player or the computer has won the game
@@ -112,22 +116,7 @@ class Board{
 
     }
 
-    public void showBoard() {
-        String separador = "-----------";
-
-        for (int rows = 0; rows < 3; rows++) {
-
-            for (int columns = 0; columns < 3; columns++) {
-
-                System.out.print(" " + board[rows][columns].getSymbol() + " ");
-
-                if (columns < 2) System.out.print("|");}
-
-            System.out.println();
-
-            if (rows < 2) {System.out.println(separador);}
-        }
-    }
+
